// src/types/anilistv2.ts
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]?: Maybe<T[SubKey]>;
};
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & {
  [SubKey in K]: Maybe<T[SubKey]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  CountryCode: any;
  FuzzyDateInt: any;
  Json: any;
};
// ...
export type ActivityReply = {
  /** The id of the parent activity */
  activityId?: Maybe<Scalars["Int"]>;
  /** The time the reply was created at */
  createdAt: Scalars["Int"];
  /** The id of the reply */
  id: Scalars["Int"];
  /** If the currently authenticated user liked the reply */
  isLiked?: Maybe<Scalars["Boolean"]>;
  /** The amount of likes the reply has */
  likeCount: Scalars["Int"];
  /** The users who liked the reply */
  likes?: Maybe<Array<Maybe<User>>>;
  /** The reply text */
  text?: Maybe<Scalars["String"]>;
  /** The user who created reply */
  user?: Maybe<User>;
  /** The id of the replies creator */
  userId?: Maybe<Scalars["Int"]>;
};
// ...
export type ActivityUnion = ListActivity | MessageActivity | TextActivity;
// ...
export type AiringSchedule = {
  /** The time the episode airs at */
  airingAt: Scalars["Int"];
  /** The airing episode number */
  episode: Scalars["Int"];
  /** The id of the airing schedule item */
  id: Scalars["Int"];
  /** The associate media of the airing episode */
  media?: Maybe<Media>;
  /** The associate media id of the airing episode */
  mediaId: Scalars["Int"];
  /** Seconds until episode starts airing */
  timeUntilAiring: Scalars["Int"];
};
// ...
export type Character = {
  /** The character's age. Note this is a string, not an int, it may contain further text and additional ages. */
  age?: Maybe<Scalars["String"]>;
  /** The characters blood type */
  bloodType?: Maybe<Scalars["String"]>;
  /** The character's birth date */
  dateOfBirth?: Maybe<FuzzyDate>;
  /** A general description of the character */
  description?: Maybe<Scalars["String"]>;
  /** The amount of user's who have favourited the character */
  favourites?: Maybe<Scalars["Int"]>;
  /** The character's gender. Usually Male, Female, or Non-binary but can be any string. */
  gender?: Maybe<Scalars["String"]>;
  /** The id of the character */
  id: Scalars["Int"];
  /** Character images */
  image?: Maybe<CharacterImage>;
  /** If the character is marked as favourite by the currently authenticated user */
  isFavourite: Scalars["Boolean"];
  /** If the character is blocked from being added to favourites */
  isFavouriteBlocked: Scalars["Boolean"];
  /** Media that includes the character */
  media?: Maybe<MediaConnection>;
  /** Notes for site moderators */
  modNotes?: Maybe<Scalars["String"]>;
  /** The names of the character */
  name?: Maybe<CharacterName>;
  /** The url for the character page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** @deprecated No data available */
  updatedAt?: Maybe<Scalars["Int"]>;
};
// ...
export type MediaList = {
  /** Map of advanced scores with name keys */
  advancedScores?: Maybe<Scalars["Json"]>;
  /** When the entry was completed by the user */
  completedAt?: Maybe<FuzzyDate>;
  /** When the entry data was created */
  createdAt?: Maybe<Scalars["Int"]>;
  /** Map of booleans for which custom lists the entry are in */
  customLists?: Maybe<Scalars["Json"]>;
  /** If the entry shown be hidden from non-custom lists */
  hiddenFromStatusLists?: Maybe<Scalars["Boolean"]>;
  /** The id of the list entry */
  id: Scalars["Int"];
  media?: Maybe<Media>;
  /** The id of the media */
  mediaId: Scalars["Int"];
  /** Text notes */
  notes?: Maybe<Scalars["String"]>;
  /** Priority of planning */
  priority?: Maybe<Scalars["Int"]>;
  /** If the entry should only be visible to authenticated user */
  private?: Maybe<Scalars["Boolean"]>;
  /** The amount of episodes/chapters consumed by the user */
  progress?: Maybe<Scalars["Int"]>;
  /** The amount of volumes read by the user */
  progressVolumes?: Maybe<Scalars["Int"]>;
  /** The amount of times the user has rewatched/read the media */
  repeat?: Maybe<Scalars["Int"]>;
  /** The score of the entry */
  score?: Maybe<Scalars["Float"]>;
  /** When the entry was started by the user */
  startedAt?: Maybe<FuzzyDate>;
  /** The watching/reading status */
  status?: Maybe<MediaListStatus>;
  /** When the entry data was last updated */
  updatedAt?: Maybe<Scalars["Int"]>;
  user?: Maybe<User>;
  /** The id of the user owner of the list entry */
  userId: Scalars["Int"];
};
// ...
export type MediaTrend = {
  /** A weighted average score of all the user's scores of the media */
  averageScore?: Maybe<Scalars["Int"]>;
  /** The day the data was recorded (timestamp) */
  date: Scalars["Int"];
  /** The episode number of the anime released on this day */
  episode?: Maybe<Scalars["Int"]>;
  /** The number of users with watching/reading the media */
  inProgress?: Maybe<Scalars["Int"]>;
  /** The related media */
  media?: Maybe<Media>;
  /** The id of the tag */
  mediaId: Scalars["Int"];
  /** The number of users with the media on their list */
  popularity?: Maybe<Scalars["Int"]>;
  /** If the media was being released at this time */
  releasing: Scalars["Boolean"];
  /** The amount of media activity on the day */
  trending: Scalars["Int"];
};
// ...
export type NotificationUnion =
  | ActivityLikeNotification
  | ActivityMentionNotification
  | ActivityMessageNotification
  | ActivityReplyLikeNotification
  | ActivityReplyNotification
  | ActivityReplySubscribedNotification
  | AiringNotification
  | FollowingNotification
  | MediaDataChangeNotification
  | MediaDeletionNotification
  | MediaMergeNotification
  | RelatedMediaAdditionNotification
  | ThreadCommentLikeNotification
  | ThreadCommentMentionNotification
  | ThreadCommentReplyNotification
  | ThreadCommentSubscribedNotification
  | ThreadLikeNotification;

/** Page of data */
export type Page = {
  activities?: Maybe<Array<Maybe<ActivityUnion>>>;
  activityReplies?: Maybe<Array<Maybe<ActivityReply>>>;
  airingSchedules?: Maybe<Array<Maybe<AiringSchedule>>>;
  characters?: Maybe<Array<Maybe<Character>>>;
  followers?: Maybe<Array<Maybe<User>>>;
  following?: Maybe<Array<Maybe<User>>>;
  likes?: Maybe<Array<Maybe<User>>>;
  media?: Maybe<Array<Maybe<Media>>>;
  mediaList?: Maybe<Array<Maybe<MediaList>>>;
  mediaTrends?: Maybe<Array<Maybe<MediaTrend>>>;
  notifications?: Maybe<Array<Maybe<NotificationUnion>>>;
  /** The pagination information */
  pageInfo?: Maybe<PageInfo>;
  recommendations?: Maybe<Array<Maybe<Recommendation>>>;
  reviews?: Maybe<Array<Maybe<Review>>>;
  staff?: Maybe<Array<Maybe<Staff>>>;
  studios?: Maybe<Array<Maybe<Studio>>>;
  threadComments?: Maybe<Array<Maybe<ThreadComment>>>;
  threads?: Maybe<Array<Maybe<Thread>>>;
  users?: Maybe<Array<Maybe<User>>>;
};
// ...
export type PageInfo = {
  /** The current page */
  currentPage?: Maybe<Scalars["Int"]>;
  /** If there is another page */
  hasNextPage?: Maybe<Scalars["Boolean"]>;
  /** The last page */
  lastPage?: Maybe<Scalars["Int"]>;
  /** The count on a page */
  perPage?: Maybe<Scalars["Int"]>;
  /** The total number of items. Note: This value is not guaranteed to be accurate, do not rely on this for logic */
  total?: Maybe<Scalars["Int"]>;
};
// ...
export type QueryThreadCommentArgs = {
  id?: InputMaybe<Scalars["Int"]>;
  sort?: InputMaybe<Array<InputMaybe<ThreadCommentSort>>>;
  threadId?: InputMaybe<Scalars["Int"]>;
  userId?: InputMaybe<Scalars["Int"]>;
};

export type QueryUserArgs = {
  id?: InputMaybe<Scalars["Int"]>;
  isModerator?: InputMaybe<Scalars["Boolean"]>;
  name?: InputMaybe<Scalars["String"]>;
  search?: InputMaybe<Scalars["String"]>;
  sort?: InputMaybe<Array<InputMaybe<UserSort>>>;
};

/** Media recommendation */
export type Recommendation = {
  /** The id of the recommendation */
  id: Scalars["Int"];
  /** The media the recommendation is from */
  media?: Maybe<Media>;
  /** The recommended media */
  mediaRecommendation?: Maybe<Media>;
  /** Users rating of the recommendation */
  rating?: Maybe<Scalars["Int"]>;
  /** The user that first created the recommendation */
  user?: Maybe<User>;
  /** The rating of the recommendation by currently authenticated user */
  userRating?: Maybe<RecommendationRating>;
};
// ...
export type Review = {
  /** The main review body text */
  body?: Maybe<Scalars["String"]>;
  /** The time of the thread creation */
  createdAt: Scalars["Int"];
  /** The id of the review */
  id: Scalars["Int"];
  /** The media the review is of */
  media?: Maybe<Media>;
  /** The id of the review's media */
  mediaId: Scalars["Int"];
  /** For which type of media the review is for */
  mediaType?: Maybe<MediaType>;
  /** If the review is not yet publicly published and is only viewable by creator */
  private?: Maybe<Scalars["Boolean"]>;
  /** The total user rating of the review */
  rating?: Maybe<Scalars["Int"]>;
  /** The amount of user ratings of the review */
  ratingAmount?: Maybe<Scalars["Int"]>;
  /** The review score of the media */
  score?: Maybe<Scalars["Int"]>;
  /** The url for the review page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** A short summary of the review */
  summary?: Maybe<Scalars["String"]>;
  /** The time of the thread last update */
  updatedAt: Scalars["Int"];
  /** The creator of the review */
  user?: Maybe<User>;
  /** The id of the review's creator */
  userId: Scalars["Int"];
  /** The rating of the review by currently authenticated user */
  userRating?: Maybe<ReviewRating>;
};
// ...
export type Studio = {
  /** The amount of user's who have favourited the studio */
  favourites?: Maybe<Scalars["Int"]>;
  /** The id of the studio */
  id: Scalars["Int"];
  /** If the studio is an animation studio or a different kind of company */
  isAnimationStudio: Scalars["Boolean"];
  /** If the studio is marked as favourite by the currently authenticated user */
  isFavourite: Scalars["Boolean"];
  /** The media the studio has worked on */
  media?: Maybe<MediaConnection>;
  /** The name of the studio */
  name: Scalars["String"];
  /** The url for the studio page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
};
// ...
export enum ThreadCommentSort {
  Id = "ID",
  IdDesc = "ID_DESC",
}
// ...
export type User = {
  /** The bio written by user (Markdown) */
  about?: Maybe<Scalars["String"]>;
  /** The user's avatar images */
  avatar?: Maybe<UserAvatar>;
  /** The user's banner images */
  bannerImage?: Maybe<Scalars["String"]>;
  bans?: Maybe<Scalars["Json"]>;
  /** When the user's account was created. (Does not exist for accounts created before 2020) */
  createdAt?: Maybe<Scalars["Int"]>;
  /** Custom donation badge text */
  donatorBadge?: Maybe<Scalars["String"]>;
  /** The donation tier of the user */
  donatorTier?: Maybe<Scalars["Int"]>;
  /** The users favourites */
  favourites?: Maybe<Favourites>;
  /** The id of the user */
  id: Scalars["Int"];
  /** If the user is blocked by the authenticated user */
  isBlocked?: Maybe<Scalars["Boolean"]>;
  /** If this user if following the authenticated user */
  isFollower?: Maybe<Scalars["Boolean"]>;
  /** If the authenticated user if following this user */
  isFollowing?: Maybe<Scalars["Boolean"]>;
  /** The user's media list options */
  mediaListOptions?: Maybe<MediaListOptions>;
  /** The user's moderator roles if they are a site moderator */
  moderatorRoles?: Maybe<Array<Maybe<ModRole>>>;
  /**
   * If the user is a moderator or data moderator
   * @deprecated Deprecated. Replaced with moderatorRoles field.
   */
  moderatorStatus?: Maybe<Scalars["String"]>;
  /** The name of the user */
  name: Scalars["String"];
  /** The user's general options */
  options?: Maybe<UserOptions>;
  /** The user's previously used names. */
  previousNames?: Maybe<Array<Maybe<UserPreviousName>>>;
  /** The url for the user page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** The users anime & manga list statistics */
  statistics?: Maybe<UserStatisticTypes>;
  /**
   * The user's statistics
   * @deprecated Deprecated. Replaced with statistics field.
   */
  stats?: Maybe<UserStats>;
  /** The number of unread notifications the user has */
  unreadNotificationCount?: Maybe<Scalars["Int"]>;
  /** When the user's data was last updated */
  updatedAt?: Maybe<Scalars["Int"]>;
};
// ...
export type UserStatisticTypes = {
  anime?: Maybe<UserStatistics>;
  manga?: Maybe<UserStatistics>;
};